packages/doeff-gemini/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doeff-gemini"
version = "0.1.0"
description = "Google Gemini integration for doeff with structured output and observability"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Kento Masui"},
]
keywords = ["gemini", "google", "effects", "structured-output", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "doeff>=0.1.0",
    "google-genai>=0.3.0",
    "google-auth>=2.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/CyberAgentAILab/doeff"
Repository = "https://github.com/CyberAgentAILab/doeff"
Issues = "https://github.com/CyberAgentAILab/doeff/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/doeff_gemini"]

[tool.uv.sources]
doeff = {workspace = true}

[dependency-groups]
dev = [
    "pillow>=11.0.0",
]
